+++ Tests/Tests/SDWebImageDownloaderTests.m
@@ -21,11 +21,6 @@
 
 @interface SDWebImageDownloader ()
 @property (strong, nonatomic, nonnull) NSOperationQueue *downloadQueue;
-
-- (nullable SDWebImageDownloadToken *)addProgressCallback:(SDWebImageDownloaderProgressBlock)progressBlock
-                                           completedBlock:(SDWebImageDownloaderCompletedBlock)completedBlock
-                                                   forURL:(nullable NSURL *)url
-                                           createCallback:(NSOperation<SDWebImageDownloaderOperation> *(^)(void))createCallback;
 @end
 
 
@@ -104,16 +99,15 @@
     [downloader invalidateSessionAndCancel:YES];
 }
 
-- (void)test07ThatAddProgressCallbackCompletedBlockWithNilURLCallsTheCompletionBlockWithNils {
+- (void)test07ThatDownloadImageWithNilURLCallsCompletionWithNils {
     XCTestExpectation *expectation = [self expectationWithDescription:@"Completion is called with nils"];
-    [[SDWebImageDownloader sharedDownloader] addProgressCallback:nil completedBlock:^(UIImage * _Nullable image, NSData * _Nullable data, NSError * _Nullable error, BOOL finished) {
-        if (!image && !data && error) {
-            [expectation fulfill];
-        } else {
-            XCTFail(@"All params except error should be nil");
-        }
-    } forURL:nil createCallback:nil];
-    [self waitForExpectationsWithTimeout:0.5 handler:nil];
+    [[SDWebImageDownloader sharedDownloader] downloadImageWithURL:nil options:0 progress:nil completed:^(UIImage * _Nullable image, NSData * _Nullable data, NSError * _Nullable error, BOOL finished) {
+        expect(image).to.beNil();
+        expect(data).to.beNil();
+        expect(error.code).equal(SDWebImageErrorInvalidURL);
+        [expectation fulfill];
+    }];
+    [self waitForExpectationsWithCommonTimeout];
 }
 
 - (void)test08ThatAHTTPAuthDownloadWorks {
@@ -263,6 +257,34 @@
         }
     }];
     [self waitForExpectationsWithCommonTimeout];
+}
+
+- (void)test17ThatMinimumProgressIntervalWorks {
+    XCTestExpectation *expectation = [self expectationWithDescription:@"Minimum progress interval"];
+    SDWebImageDownloaderConfig *config = SDWebImageDownloaderConfig.defaultDownloaderConfig;
+    config.minimumProgressInterval = 0.51; // This will make the progress only callback once
+    SDWebImageDownloader *downloader = [[SDWebImageDownloader alloc] initWithConfig:config];
+    NSURL *imageURL = [NSURL URLWithString:@"http://www.ioncannon.net/wp-content/uploads/2011/06/test2.webp"];
+    __block NSUInteger allProgressCount = 0; // All progress (including operation start / first HTTP response, etc)
+    __block NSUInteger validProgressCount = 0; // Only progress from `URLSession:dataTask:didReceiveData:`
+    [downloader downloadImageWithURL:imageURL options:0 progress:^(NSInteger receivedSize, NSInteger expectedSize, NSURL * _Nullable targetURL) {
+        allProgressCount++;
+        if (expectedSize <= 0 || receivedSize <= 0) {
+            // ignore the progress callback until we receive data
+            return;
+        }
+        validProgressCount++;
+    } completed:^(UIImage * _Nullable image, NSData * _Nullable data, NSError * _Nullable error, BOOL finished) {
+        if (allProgressCount > 1 && validProgressCount == 1) {
+            [expectation fulfill];
+        } else {
+            XCTFail(@"Progress callback more than once");
+        }
+    }];
+     
+    [self waitForExpectationsWithCommonTimeoutUsingHandler:^(NSError * _Nullable error) {
+        [downloader invalidateSessionAndCancel:YES];
+    }];
 }
 
 /**